Extract pygame event pump
